// src/common/health/health-check.interface.ts
export enum HealthStatus {
  HEALTHY = 'healthy',
  UNHEALTHY = 'unhealthy',
  DEGRADED = 'degraded',
}

export interface ServiceHealth {
  name: string;
  url: string;
  status: HealthStatus;
  responseTime: number; // in milliseconds
  lastChecked: Date;
  error?: Error; // optional error message if the service is unhealthy
}