// src/common/health/health-check.service.ts
import { Injectable, Logger } from '@nestjs/common';
import { HealthStatus, type ServiceHealth } from './health-check.interface';
import { HttpService } from '@nestjs/axios';
import { CircuitBreakerService } from '../circuit-breaker/circuit-breaker.service';
import { serviceConfig } from 'src/config/gateway.config';
import { firstValueFrom, timeout } from 'rxjs';

@Injectable()
export class HealthCheckService {
  private readonly logger = new Logger(HealthCheckService.name);
  private readonly healthCache = new Map<string, ServiceHealth>();

  constructor(
    private readonly httpService: HttpService,
    private readonly circuitBreakerService: CircuitBreakerService,
  ) {}

  async checkServiceHealth(
    serviceName: keyof typeof serviceConfig,
  ): Promise<ServiceHealth> {
    const service = serviceConfig[serviceName];
    const startTime = Date.now();

    try {
      await this.circuitBreakerService.executeWithCircuitBreaker(
        async () => {
          const response = await firstValueFrom(
            this.httpService
              .get(`${service.url}/health`, {
                timeout: service.timeout,
              })
              .pipe(timeout(service.timeout)),
          );

          return response.status;
        },
        `health-${serviceName}`,
        () => {
          throw new Error('Circuit breaker fallback: Service is unhealthy');
        },
      );

      const responseTime = Date.now() - startTime;
      const serviceHealth: ServiceHealth = {
        name: serviceName,
        url: service.url,
        status: HealthStatus.HEALTHY,
        responseTime,
        lastChecked: new Date(),
      };
      this.healthCache.set(serviceName, serviceHealth);
      return serviceHealth;
    } catch (error: unknown) {
      const responseTime = Date.now() - startTime;

      const serviceHealth: ServiceHealth = {
        name: serviceName,
        url: service.url,
        status: HealthStatus.UNHEALTHY,
        responseTime,
        lastChecked: new Date(),
      };

      const errorMessage =
        error instanceof Error ? error.message : 'Unknown error';

      serviceHealth.error = new Error(errorMessage);
      this.healthCache.set(serviceName, serviceHealth);
      this.logger.error(`Health check failed for ${serviceName}`, errorMessage);

      return serviceHealth;
    }
  }

  async checkAllServices(): Promise<ServiceHealth[]> {
    const services: (keyof typeof serviceConfig)[] = [
      'users',
      'products',
      'checkout',
      'payments',
    ];

    const healthChecks = await Promise.allSettled(
      services.map((serviceName) => this.checkServiceHealth(serviceName)),
    );

    return healthChecks.map((result, index) => {
      if (result.status === 'fulfilled') {
        return result.value;
      } else {
        return {
          name: services[index],
          url: serviceConfig[services[index]].url,
          status: HealthStatus.UNHEALTHY as const,
          responseTime: 0,
          lastChecked: new Date(),
          error:
            result.reason instanceof Error
              ? result.reason
              : new Error('Unknown error'),
        };
      }
    });
  }

  getCacheHealth(serviceName: string): ServiceHealth | undefined {
    return this.healthCache.get(serviceName);
  }

  getAllCacheHealth(): ServiceHealth[] {
    return Array.from(this.healthCache.values());
  }
}
